refactor: clean up shim system and add wp_timezone() shim
