+++ src/Framework/PaymentGateways/Routes/RouteSignature.php
@@ -80,7 +80,6 @@ class RouteSignature
      */
     public function createExpirationTimestamp()
     {
-        Shim::load( 'current_datetime' );
         return (string)current_datetime()->modify('+1 day')->getTimestamp();
     }
